+++ src/controllers/task.controller.js
@@ -1,13 +1,14 @@
 const { v4: uuidv4 } = require("uuid");
 const { pool } = require("../db/connection");
-const {tasksDecorator} = require("../decorators/task.decorator");
+const { tasksDecorator } = require("../decorators/task.decorator");
 
 const store = async (req, res) => {
-  const { title, description, status, category_id, user_id, tags } = req.body;
+  const { title, description, status, category_id, tags } = req.body;
+  const user_id = req.user.id;
 
-  if (!title || !user_id) {
+  if (!title) {
     return res.status(400).json({
-      message: "title y user_id son obligatorios"
+      message: "title es obligatorio"
     });
   }
 
@@ -27,7 +28,7 @@ const store = async (req, res) => {
     user_id
   ]);
 
-  /* guardar relacion con tags */
+  /* guardar relación con tags */
   if (tags && tags.length > 0) {
     for (const tagId of tags) {
       await pool.query(
@@ -40,16 +41,10 @@ const store = async (req, res) => {
   res.status(201).json({
     message: "Tarea creada correctamente"
   });
-}
+};
 
 const index = async (req, res) => {
-  const { user_id } = req.query;
-
-  if (!user_id) {
-    return res.status(400).json({
-      message: "user_id es obligatorio"
-    });
-  }
+  const user_id = req.user.id;
 
   const [rows] = await pool.query(
     `
@@ -69,11 +64,12 @@ const index = async (req, res) => {
   );
 
   res.json(tasksDecorator(rows));
-}
+};
 
 const update = async (req, res) => {
   const { id } = req.params;
   const { title, description, status, category_id, tags } = req.body;
+  const user_id = req.user.id;
 
   const query = `
     UPDATE tasks
@@ -82,7 +78,7 @@ const update = async (req, res) => {
       description = COALESCE(?, description),
       status = COALESCE(?, status),
       category_id = COALESCE(?, category_id)
-    WHERE id = ?
+    WHERE id = ? AND user_id = ?
   `;
 
   await pool.query(query, [
@@ -90,7 +86,8 @@ const update = async (req, res) => {
     description,
     status,
     category_id,
-    id
+    id,
+    user_id
   ]);
 
   if (tags !== undefined) {
@@ -111,20 +108,27 @@ const update = async (req, res) => {
   res.json({
     message: "Tarea actualizada correctamente"
   });
-}
+};
 
 const destroy = async (req, res) => {
   const { id } = req.params;
+  const user_id = req.user.id;
+  
+  /* eliminar relaciones con tags */
+  await pool.query(
+    `DELETE FROM tags_tasks WHERE task_id = ?`,
+    [id]
+  );
 
   await pool.query(
-    "DELETE FROM tasks WHERE id = ?",
-    [id]
+    "DELETE FROM tasks WHERE id = ? AND user_id = ?",
+    [id, user_id]
   );
 
   res.json({
     message: "Tarea eliminada"
   });
-}
+};
 
 module.exports = {
   store,
